Configured So Device Name is used instead of UDID
I added in some code to make the folders in which the information is
stored have their names be the device name, not the UDID to make things
simpler. If you do need the UDID, it is stored in a document within the
backup. Backups are also called Encrypted Backup and Decrypted backup
instead of the device UDID for the encrypted and the name _unback_ for
the decrypted.